fix problems in SMURFParser and use utils:ConfigMaps
In order to allow switching from utils::ConfigMaps to external
ConfigMaps on a separate branch, it makes life easier - and
possible to revert the whole change - if it is done on a separate
branch. Doing so before merging factory_load_interface as a branch
could create quite a marathon of merging problems.