Add function getAbsoluteFilePaths
This function was created by Behnam Asadi and merged manually  with some
modifications. It extends the smurf parser's functionality by providing
the paths of all files in the smurf hierarchy on top of the data
contained within them, which allows external tools to read in and work
with the files.